% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelPrint-package.R
\docType{package}
\name{accelPrint-package}
\alias{accelPrint}
\alias{accelPrint-package}
\title{accelPrint: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
