% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fingerprints.R
\name{plot_fingerprints}
\alias{plot_fingerprints}
\title{Plot fingerprints from densities in accelerometry data}
\usage{
plot_fingerprints(data, lags, sample_rate = NULL, max_vm = 3)
}
\arguments{
\item{data}{A data frame with accelerometry values (e.g., x, y, z, and time), preferably from walking}

\item{lags}{vector of lag values in seconds}

\item{sample_rate}{optional specification of sample rate (samples per sec), if not specified will be inferred}

\item{max_vm}{maximum vector magnitude for plotting, defaults to 3}
}
\value{
A plot of the fingerprints for each provided lag across all provided seconds
}
\description{
This function takes in walking data and lags and plots "fingerprints"
}
\examples{
# Load example walking bout data
data(walking_data)

# Run the function
plot_fingerprints(walking_data, lags = c(0.15, 0.30, 0.45))
}
