% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_accel}
\alias{raw_accel}
\title{Example accelerometry dataset}
\format{
A data frame with N rows and M columns:
\describe{
\item{time}{POSIXct. Time stamp.}
\item{x}{Numeric. X-axis value in g}
\item{y}{Numeric. Y-axis value in g}
\item{z}{Numeric. Z-axis value in g}
}
}
\source{
Karas, M., Urbanek, J., Crainiceanu, C., Harezlak, J., & Fadel, W. (2021). Labeled raw accelerometry data captured during walking, stair climbing and driving (version 1.0.0). PhysioNet. https://doi.org/10.13026/51h0-a262.
}
\usage{
raw_accel
}
\description{
A dataset included in accelPrint to illustrate usage.
Raw 100 Hz accelerometry data from one subject is downloaded from Physionet https://physionet.org/content/accelerometry-walk-climb-drive/1.0.0/raw_accelerometry_data/
The first 5 minutes are provided as an example
}
\keyword{datasets}
