% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_df}
\alias{sample_df}
\title{Example sample dataset}
\format{
A data frame with N rows and M columns:
\describe{
\item{x}{Numeric. X-axis value.}
\item{y}{Numeric. Y-axis value.}
\item{z}{Numeric. Z-axis value.}
\item{time}{POSIXct. Time stamp.}
}
}
\source{
Simulated data or citation if applicable
}
\usage{
sample_df
}
\description{
A dataset included in accelPrint to illustrate usage.
}
\keyword{datasets}
