% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walking_data}
\alias{walking_data}
\title{Example walking accelerometry dataset}
\format{
A data frame with N rows and M columns:
\describe{
\item{time}{POSIXct. Time stamp.}
\item{x}{Numeric. X-axis value in g}
\item{y}{Numeric. Y-axis value in g}
\item{z}{Numeric. Z-axis value in g}
\item{second}{POSIXct. Time stamp in seconds}
\item{bout_seconds}{Numeric, number of seconds in walking bout}
}
}
\source{
Karas, M., Urbanek, J., Crainiceanu, C., Harezlak, J., & Fadel, W. (2021). Labeled raw accelerometry data captured during walking, stair climbing and driving (version 1.0.0). PhysioNet. https://doi.org/10.13026/51h0-a262.
}
\usage{
walking_data
}
\description{
A dataset included in accelPrint to illustrate usage.
This dataset is the result of running the get_walking function on raw_accel
}
\keyword{datasets}
