% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelPrint-package.R
\docType{package}
\name{accelPrint-package}
\alias{accelPrint}
\alias{accelPrint-package}
\title{accelPrint: Fingerprinting with Gait Characteristics to Identify Participants}
\description{
Creates a waking "fingerprint" that can identify a participant, taking in raw, sub-second level data, defining walking, and then converting that to a fingerprint.
}
\author{
\strong{Maintainer}: Lily Koffman \email{lkoffma2@jhmi.edu}

Authors:
\itemize{
  \item John Muschelli \email{muschellij2@gmail.com}
}

}
\keyword{internal}
